Fixed some warnings that annoyed me, the one in RenderTargetTexture could be important.
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@2795 610dfd11-2700-0410-9b68-9ae28eac5983

+++ src/Scene.cpp
@@ -186,7 +186,7 @@ void Scene::idle() {
     H3DTime sleep_time= min_frame_time-dt0;
     if ( sleep_time > 0 ) {
 #ifdef H3D_WINDOWS
-      Sleep ( sleep_time*1000.0 );
+      Sleep ( (DWORD)(sleep_time*1000.0) );
 #else
       usleep( sleep_time*1e6 );
 #endif
